# docs/backend/openapi.v1.yaml
...
openapi: 3.1.0
info:
  title: LabelScan API
  version: v1
  description: Generated inventory of the currently implemented LabelScan API. Production disables the
    interactive OpenAPI and Swagger endpoints; this reviewed artifact defines pentest scope.
paths:
  /v1/auth/login:
    post:
      summary: Login
      operationId: login_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Request validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security: []
  /v1/o/{organization_slug}/auth/login:
    post:
      summary: Organization Login
      operationId: organization_login_v1_o__organization_slug__auth_login_post
      parameters:
      - name: organization_slug
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 63
          pattern: ^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$
          title: Organization Slug
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Request validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security: []
  /v1/mobile/auth/login:
    post:
      summary: Mobile Login
      operationId: mobile_login_v1_mobile_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileLoginResponse'
        '400':
          description: Request validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security: []
  /v1/mobile/auth/refresh:
    post:
      summary: Mobile Refresh
      operationId: mobile_refresh_v1_mobile_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileLoginResponse'
        '400':
          description: Request validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security: []
  /v1/mobile/auth/logout:
    post:
      summary: Mobile Logout
      operationId: mobile_logout_v1_mobile_auth_logout_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '204':
          description: Successful Response
        '400':
          description: Request validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security: []
  /v1/auth/refresh:
    post:
      summary: Browser Refresh
      operationId: browser_refresh_v1_auth_refresh_post
      parameters:
      - name: labelscan_refresh
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Labelscan Refresh
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Request validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security: []
  /v1/auth/logout:
    post:
      summary: Browser Logout
      operationId: browser_logout_v1_auth_logout_post
      parameters:
      - name: labelscan_refresh
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Labelscan Refresh
      responses:
        '204':
          description: Successful Response
        '400':
          description: Request validation failed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security: []
...
components:
  schemas:
...
    CurrentUserResponse:
      properties:
        id:
          type: string
          title: Id
        username:
          type: string
          title: Username
        display_name:
          type: string
          title: Display Name
        role:
          type: string
          title: Role
        store_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Store Code
        organization_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Organization Id
        organization_slug:
          type: string
          title: Organization Slug
          default: labelscan
        store_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Store Id
        business_portal_ids:
          items:
            type: string
          type: array
          title: Business Portal Ids
        business_portal_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Business Portal Id
        trade_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Trade Code
        client_type:
          type: string
          title: Client Type
      type: object
      required:
      - id
      - username
      - display_name
      - role
      - store_code
      - client_type
      title: CurrentUserResponse
...
    LoginRequest:
      properties:
        username:
          type: string
          maxLength: 254
          minLength: 1
          title: Username
        password:
          type: string
          maxLength: 128
          minLength: 1
          title: Password
      additionalProperties: false
      type: object
      required:
      - username
      - password
      title: LoginRequest
    LoginResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
        user:
          $ref: '#/components/schemas/CurrentUserResponse'
      type: object
      required:
      - access_token
      - expires_in
      - user
      title: LoginResponse
...
    MobileLoginResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
        user:
          $ref: '#/components/schemas/CurrentUserResponse'
        refresh_token:
          type: string
          title: Refresh Token
        refresh_expires_in:
          type: integer
          title: Refresh Expires In
      type: object
      required:
      - access_token
      - expires_in
      - user
      - refresh_token
      - refresh_expires_in
      title: MobileLoginResponse
...
    RefreshRequest:
      properties:
        refresh_token:
          type: string
          maxLength: 256
          minLength: 32
          title: Refresh Token
      additionalProperties: false
      type: object
      required:
      - refresh_token
      title: RefreshRequest
...
    Problem:
      type: object
      description: Sanitized RFC 9457-style error envelope.
      required:
      - type
      - title
      - status
      - detail
      - error_code
      - timestamp
      - retriable
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        error_code:
          type: string
        correlation_id:
          anyOf:
          - type: string
          - type: 'null'
        trace_id:
          anyOf:
          - type: string
          - type: 'null'
        timestamp:
          type: string
          format: date-time
        retriable:
          type: boolean
        errors:
          type: array
          items:
            type: object
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 15-minute JWT with jti, sid, issuer, audience, tenant, role, and scopes
servers:
- url: /
  description: same-origin deployment